requirements.txt
atlassian-python-api>=3.35.0
GitPython>=3.1.24